Reorganize cli to modules
